% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFullscreen.R
\name{addFullscreen}
\alias{addFullscreen}
\title{Full screen button for Leaflet}
\usage{
addFullscreen(map)
}
\arguments{
\item{map}{a Leaflet for R map object}
}
\value{
The modified map object
}
\description{
Plugin for Leaflet that shows a button to show the map on the full screen
}
\details{
No frills, no options. This function displays a full screen button under the + and - buttons,
and fills the screen with your map when the user clicks it.

This code is an R wrapper for JavaScript package Leaflet.fullscreen, and this package
includes the .js, .css, and .jvg files from \url{https://github.com/runette/Leaflet.fullscreen}.
Leaflet.fullscreen is Licensed under ISC.
}
\section{Author}{

Bradley W. Compton \href{mailto:bcompton@umass.edu}{bcompton@umass.edu}
}

\examples{
require(leaflet.lagniappe)
require(leaflet)
leaflet() |>
addTiles() |>
addFullscreen()

}
