% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmGeocoder.R
\name{osmGeocoder}
\alias{osmGeocoder}
\title{Open Street Map Geocoding for Leaflet}
\usage{
osmGeocoder(
  map,
  collapsed = "false",
  position = "topright",
  text = "Locate",
  placeholder = "Type an address...",
  bounds = "null",
  email = ""
)
}
\arguments{
\item{map}{a Leaflet for R map object}

\item{collapsed}{whether text window is collapsed or not}

\item{position}{the position of the control}

\item{text}{the text of the submit button}

\item{placeholder}{the text of the search input placeholder}

\item{bounds}{a L.LatLngBounds object to limit the results to}

\item{email}{an email string with a contact to provide to Nominatim. Useful if
you are doing lots of queries}
}
\value{
The modified map object
}
\description{
Plugin for Leaflet that allows users to enter a place name or address to geolocate a map
}
\details{
This plugin uses Open Street Maps' Nominatim geocoder to resolve an address and locate
the Leaflet map. There are several arguments to fine-tune the text box. These are JavaScript
arguments, so their format is different from R. Collapsed should be a text string, 'true' or
'false'. The others are all text strings.

Note: there are limits to the free Nominatim geocoder (e.g., no more than 1 query/sec)
See \url{https://operations.osmfoundation.org/policies/nominatim/} for details.

This code is an R wrapper for JavaScript package leaflet-control-osm-geocoder, and this package
includes the .js and .css files from \url{https://github.com/k4r573n/leaflet-control-osm-geocoder}.
leaflet-control-osm-geocoder is licensed under the BSD 2-Clause "Simplified" License.
}
\section{Author}{

Bradley W. Compton \href{mailto:bcompton@umass.edu}{bcompton@umass.edu}
}

\examples{
require(leaflet.lagniappe)
require(leaflet)
leaflet() |>
addTiles() |>
osmGeocoder()

}
